+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+- feat: Campaign location showing city or state based on the office being run for
+
 ## v3.4.29 - 2026-05-18
 
 - fix: CSP `connect-src` missing `googletagmanager.com`, blocking GTM debug requests